+++ src/jmodem/Sampler.java
@@ -8,32 +8,32 @@ public class Sampler implements InputStream {
 	public final int resolution = 1024;
 
 	private final int N = width * resolution;
-	private final float[][] filt;
+	private final double[][] filt;
 	private final int coeffs_len;
-	private float time;
-	private float freq;
-	private final float[] buff;
+	private double time;
+	private double freq;
+	private final double[] buff;
 	private int index;
 	private final InputStream src;
 
 	public Sampler(InputStream source) {
 		src = source;
-		float[] h = new float[2 * N];
+		double[] h = new double[2 * N];
 		for (int i = -N; i < N; i++) {
 			double u = i;
 			double cosine = Math.cos(0.5 * Math.PI * u / N);
 			double window = cosine * cosine;
-			h[N + i] = (float) (sinc(Math.PI * u / resolution) * window);
+			h[N + i] = sinc(Math.PI * u / resolution) * window;
 		}
 		coeffs_len = 2 * width;
-		filt = new float[resolution][coeffs_len];
+		filt = new double[resolution][coeffs_len];
 		for (int i = 0; i < resolution; i++) {
 			for (int j = 0; j < coeffs_len; j++) {
 				filt[i][j] = h[i + (coeffs_len - 1 - j) * resolution];
 			}
 		}
 
-		buff = new float[coeffs_len]; // zeroes (by default)
+		buff = new double[coeffs_len]; // zeroes (by default)
 		index = width;
 		time = width + 1;
 		freq = 1f;
@@ -44,11 +44,11 @@ public class Sampler implements InputStream {
 	}
 
 	@Override
-	public void read(float[] buffer, int offset, int size) throws IOException {
+	public void read(double[] buffer, int offset, int size) throws IOException {
 		for (int o = offset; o < size; o++) {
 			int k = (int) time;
 			int j = (int) ((time - k) * resolution);
-			float[] coeffs = filt[j];
+			double[] coeffs = filt[j];
 			for (int end = k + width; index < end; index++) {
 				System.arraycopy(buff, 1, buff, 0, buff.length - 1);
 				src.read(buff, buff.length - 1, 1); // push to buff's end
@@ -56,17 +56,17 @@ public class Sampler implements InputStream {
 			time += freq;
 			double result = 0f;
 			for (int i = 0; i < buff.length; i++) {
-				result += (coeffs[i] * (double) buff[i]);
+				result += (coeffs[i] * buff[i]);
 			}
-			buffer[o] = (float) result;
+			buffer[o] = result;
 		}
 	}
 
-	public void updateTime(float dt) {
+	public void updateTime(double dt) {
 		time += dt;
 	}
 
-	public void updateFreq(float df) {
+	public void updateFreq(double df) {
 		freq += df;
 	}
